# docs/swagger.yaml
definitions:
...
  request.InstanceCreateRequest:
    properties:
      challenge_id:
        type: string
    required:
    - challenge_id
    type: object
  request.InstanceRemoveRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  request.InstanceRenewRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  request.SubmissionCreateRequest:
    properties:
      challenge_id:
        description: 题目 Id
        type: string
      flag:
        description: 提交内容
        type: string
      game_id:
        description: 比赛 Id
        type: integer
      team_id:
        description: 团队 Id
        type: string
    type: object
  request.TeamCreateRequest:
    properties:
      captain_id:
        type: string
      name:
        type: string
    required:
    - captain_id
    - name
    type: object
  request.TeamDeleteRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  request.TeamJoinRequest:
    properties:
      team_id:
        type: string
      user_id:
        type: string
    required:
    - team_id
    - user_id
    type: object
  request.TeamQuitRequest:
    properties:
      team_id:
        type: string
      user_id:
        type: string
    required:
    - team_id
    - user_id
    type: object
  request.TeamUpdateRequest:
    properties:
      captain_id:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - captain_id
    - id
    - name
    type: object
...
info:
  contact: {}
  description: 没有其他东西啦，仅仅是所有的后端接口，不要乱用哦
  title: PgsHub Backend API
  version: "1.0"
paths:
...
  /api/instances/:
    delete:
      description: 停止并删除容器
      parameters:
      - description: InstanceRemoveRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.InstanceRemoveRequest'
      produces:
      - application/json
      responses: {}
      summary: 停止并删除容器
      tags:
      - 实例
    get:
      description: 实例全部查询
      produces:
      - application/json
      responses: {}
      summary: 实例全部查询
      tags:
      - 实例
    post:
      consumes:
      - application/json
      description: 创建实例
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: InstanceCreateRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.InstanceCreateRequest'
      produces:
      - application/json
      responses: {}
      summary: 创建实例
      tags:
      - 实例
    put:
      description: 容器续期
      parameters:
      - description: InstanceRenewRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.InstanceRenewRequest'
      produces:
      - application/json
      responses: {}
      summary: 容器续期
      tags:
      - 实例
  /api/instances/{id}:
    get:
      description: 实例查询
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 实例查询
      tags:
      - 实例
  /api/submissions/:
    get:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: 题目 Id
        in: query
        name: challenge_id
        type: string
      - description: 比赛 Id
        in: query
        name: game_id
        type: integer
      - description: 是否详细
        in: query
        name: is_detailed
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: size
        type: integer
      - description: 评判结果
        in: query
        name: status
        type: integer
      - description: 团队 Id
        in: query
        name: team_id
        type: string
      - description: 用户 Id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses: {}
      summary: 提交记录查询
      tags:
      - 提交
    post:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: SubmissionCreateRequest
        in: body
        name: 创建请求
        required: true
        schema:
          $ref: '#/definitions/request.SubmissionCreateRequest'
      produces:
      - application/json
      responses: {}
      summary: 提交
      tags:
      - 提交
  /api/teams/:
    delete:
      consumes:
      - application/json
      description: 删除团队
      parameters:
      - description: TeamDeleteRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.TeamDeleteRequest'
      produces:
      - application/json
      responses: {}
      summary: 删除团队
      tags:
      - 团队
    get:
      consumes:
      - application/json
      description: 查找团队
      parameters:
      - in: query
        name: captain_id
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 查找团队
      tags:
      - 团队
    post:
      consumes:
      - application/json
      description: 创建团队
      parameters:
      - description: TeamCreateRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.TeamCreateRequest'
      produces:
      - application/json
      responses: {}
      summary: 创建团队
      tags:
      - 团队
    put:
      consumes:
      - application/json
      description: 更新团队
      parameters:
      - description: TeamUpdateRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.TeamUpdateRequest'
      produces:
      - application/json
      responses: {}
      summary: 更新团队
      tags:
      - 团队
  /api/teams/id/{id}:
    get:
      consumes:
      - application/json
      description: 查找团队
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 查找团队
      tags:
      - 团队
  /api/teams/members/:
    delete:
      consumes:
      - application/json
      description: 退出团队
      parameters:
      - description: TeamQuitRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.TeamQuitRequest'
      produces:
      - application/json
      responses: {}
      summary: 退出团队
      tags:
      - 团队
    post:
      consumes:
      - application/json
      description: 加入团队
      parameters:
      - description: TeamJoinRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.TeamJoinRequest'
      produces:
      - application/json
      responses: {}
      summary: 加入团队
      tags:
      - 团队
...
swagger: "2.0"